degrade archlinux support to "best-effort"
recently, the Archlinux docker image used for testing has started to fail
in CI tests, giving us failures that are hard to troubleshoot.
This change degrades Archlinux support to "best-effort", meaning that we
don't explicitly run CI against it.
As the ArchLinux code is very minimal anyways, this should be
a reasonable compromise