bump ansible to 2.17, raise ninimum python to 3.7 (#200)
bump ansible+python to 2.17+3.7, update distros

with the ansible-2.17 release, python 2 and 3.6 have been removed
from the officially supported list. We therefore have to upgrade
our minimum tested version to 3.7, which is a breaking change
as it requires removing official support for RockyLinux 8.

Also use this opportunity to add support for Ubuntu 24.04.